requirements.txt
flask==2.3.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.2.0
numpy==1.24.3
scikit-learn==1.3.0